Revert "PyCrypto requires RSA values to be long"
This reverts commit a1703e15d4e9f1afc00a746cac1d1df585bfb546.

+++ DeDRM_plugin/ineptepub.py
@@ -311,7 +311,7 @@ def _load_crypto_pycrypto():
             total = 0
             for byte in bytes:
                 total = (total << 8) + byte
-            return long(total)
+            return total
 
         def decrypt(self, data):
             return _PKCS1_v1_5.new(self._rsa).decrypt(data, 172)
